server/hywz-flask/requirements.txt
aniso8601==3.0.2
astunparse==1.5.0
awsebcli==3.14.3
backcall==0.1.0
blinker==1.4
boto3==1.7.80
botocore==1.10.80
cachetools==2.0.1
cement==2.8.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
colour==0.1.5
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.7.1
Django==2.0.3
django-allauth==0.35.0
django-rest-auth==0.9.3
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
Flask==1.0.2
Flask-RESTful==0.3.6
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
geojson==2.4.0
google-api-core==0.1.4
google-auth==1.4.1
google-cloud==0.32.0
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.1
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.11.0
httplib2==0.11.3
idna==2.6
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
lxml==4.2.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mongoengine==0.15.0
mpld3==0.3
numpy==1.14.0
oauth2client==3.0.0
oauthlib==2.0.7
pandas==0.22.0
parso==0.3.1
pathspec==0.5.5
pickleshare==0.7.4
Pillow==5.1.0
pixiedust==1.1.11
pixiedust-node==0.2.5
ply==3.8
prompt-toolkit==1.0.15
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.2.post1
psutil==5.4.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
PyJWT==1.6.1
pymongo==3.6.1
pyparsing==2.2.0
pyreadline==2.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
xlrd==1.1.0
